extract emojis to constants
