+++ packages/core/src/constants.ts
@@ -62,3 +62,7 @@ export const OLLAMA_API_BASE = "http://localhost:11434/v1"
 export const PROMPTFOO_CACHE_PATH = ".genaiscript/cache/tests"
 export const PROMPTFOO_CONFIG_DIR = ".genaiscript/config/tests"
 export const PROMPTFOO_REMOTE_API_PORT = 15500
+
+export const EMOJI_SUCCESS = "✅"
+export const EMOJI_FAIL = "❌"
+export const EMOJI_UNDEFINED = "?"
